+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Added details when hovering over dependencies
+
 ## `[0.3.1]` - July 17, 2022
 
 ### Added